requirements.txt

streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
textblob
python-dotenv
requests
shap
